[WiP] Lottie Front Page
This took some time as there are some issues with the library. If you
put the animation in reverse then the count doesn't work, even if you
try to use the `loop` event, so it took a lot of time and debugging to
find a solution which works in all 6 use-cases.